+++ apps/router-e2e/__e2e__/native-navigation/app/index.tsx
@@ -34,6 +34,7 @@ const HomeIndex = () => {
       <CaseLink href="/js-tabs" text="JS Tabs" />
       <CaseLink href="/top-tabs" text="JS Top Tabs" />
       <CaseLink href="/experimental-stack" text="Experimental Stack" />
+      <CaseLink href="/color" text="Color" />
     </ScrollView>
   );
 };